Bounds on the strength of unobserved confounders using observed covariates

Bounds on the strength of unobserved confounders using observed covariates, as in Cinelli and Hazlett (2020).

ovb_bounds(...)

# S3 method for lm
ovb_bounds(
  model,
  treatment,
  benchmark_covariates,
  kd = 1,
  ky = kd,
  reduce = TRUE,
  bound = c("partial r2", "partial r2 no D", "total r2"),
  adjusted_estimates = TRUE,
  alpha = 0.05,
  ...
)

Arguments

...
arguments passed to other methods. First argument should either be an lm model with the outcome regression, or a formula describing the model along with the data.frame containing the variables of the model.

model
An lm object with the outcome regression.

treatment
A character vector with the name of the treatment variable of the model.

benchmark_covariates
a character vector of the names of covariates that will be used to bound the plausible strength of the unobserved confounders.

kd
numeric vector. Parameterizes how many times stronger the confounder is related to the treatment in comparison to the observed benchmark covariate. Default value is 1 (confounder is as strong as benchmark covariate).

ky
numeric vector. Parameterizes how many times stronger the confounder is related to the outcome in comparison to the observed benchmark covariate. Default value is the same as kd.

reduce
Should the bias adjustment reduce or increase the absolute value of the estimated coefficient? Default is TRUE.

bound
type of bounding procedure. Currently only "partial r2" is implemented.

adjusted_estimates
should the bounder also compute the adjusted estimates?

alpha
significance level for computing the adjusted confidence intervals. Default is TRUE.

Value

The function returns a data.frame with the bounds on the strength of the unobserved confounder as well with the adjusted point estimates, standard errors and t-values (optional, controlled by argument adjusted_estimates).

Details

Currently it implements only the bounds based on partial R2. Other bounds will be implemented soon.

References

Cinelli, C. and Hazlett, C. (2020), "Making Sense of Sensitivity: Extending Omitted Variable Bias." Journal of the Royal Statistical Society, Series B (Statistical Methodology).

Examples

# runs regression model
model <- lm(peacefactor ~ directlyharmed + age + farmer_dar + herder_dar +
             pastvoted + hhsize_darfur + female + village, data = darfur)

# bounds on the strength of confounders 1, 2, or 3 times as strong as female
# and 1,2, or 3 times as strong as pastvoted
ovb_bounds(model, treatment = "directlyharmed",
          benchmark_covariates = c("female", "pastvoted"),
          kd = 1:3)

#>    bound_label      r2dz.x     r2yz.dx adjusted_estimate adjusted_se adjusted_t
#> 1    1x female 0.009164287 0.124640923        0.07522027  0.02187333   3.438904
#> 2    2x female 0.018328573 0.249324064        0.05291517  0.02035006   2.600246
#> 3    3x female 0.027492860 0.374050471        0.03039602  0.01867006   1.628062
#> 4 1x pastvoted 0.001491963 0.005109456        0.09551771  0.02322921   4.111965
#> 5 2x pastvoted 0.002983926 0.010218957        0.09371690  0.02318681   4.041819
#> 6 3x pastvoted 0.004475890 0.015328504        0.09191338  0.02314421   3.971334
#>   adjusted_lower_CI adjusted_upper_CI
#> 1       0.032282966        0.11815758
#> 2       0.012968035        0.09286231
#> 3      -0.006253282        0.06704533
#> 4       0.049918808        0.14111661
#> 5       0.048201225        0.13923257
#> 6       0.046481338        0.13734543