Sensitivity analysis plots for sensemakr

This function provides the contour and extreme scenario sensitivity plots of the sensitivity analysis results obtained with the function sensemakr. They are basically dispatchers to the core plot functions ovb_contour_plot and ovb_extreme_plot.

# S3 method for sensemakr
plot(x, type = c("contour", "extreme"), ...)

Arguments

x
an object of class sensemakr.

type
type of sensitivity plot. It can be "contour", for contour plots of omitted variable bias as in ovb_contour_plot; or, "extreme" for extreme scenarios plots of omitted variable bias as in ovb_extreme_plot.

...
arguments passed to the plot functions. Check arguments in ovb_contour_plot and ovb_extreme_plot.