Creates orthogonal residuals

This function is an auxiliary function for simulation purposes. It creates a vector of n standard normal random variables, residualizes this vector against a matrix of covariates C, then standardizes the vector again.

resid_maker(n, C)

Arguments

n

sample size.

C

a numeric matrix with n rows and p columns.

Value

The function returns a numeric vector orthogonal to C.